[all][l]: remove constituent financials no longer working (yahoo finance stopped) and we have a new version in separate repo here https://github.com/datasets/s-and-p-companies-financials - fixes #19.
* requirements.txt: also update reflecting refactors over a while
* Makefile: update to reflect removal (needs more work to be up to date)

As @noahg wrote in original issue:

Hello, cool project. I propose that the constituent-financials be
removed from this data package.

The constituents.csv table is useful in its own right, containing
symbol, name, and sector. This file currently still builds but the
overall makefile is failing for multiple reasons, one of which is that
the Yahoo Finance API [seems to have have changed its terms or perhaps
is going away
altogether.](https://stackoverflow.com/questions/44057580/yahoo-finance-api-changes-2017)
which is causing the constituents-financials to fail.

My assumption is that stock quote and other financial data will forever
be relegated to depending on free services that should be considered at
a minimum as less dependable than the Wikipedia dependence for just the
constituents data.

The financials are useful but they could be appended by a separate data
package that validates the financials only.